+++ docs/src/app/docs/layout.tsx
@@ -1,12 +1,11 @@
-import { HeaderNav } from '../../components/navigation/header-nav';
 import '../../styles/global.css';
-import keystaticConfig from '../../../keystatic.config';
-import { createReader } from '@keystatic/core/reader';
 import { SideNav } from '../../components/navigation/side-nav';
 import { NavGroup } from '../../components/navigation/nav-group';
 import { NavItem } from '../../components/navigation/nav-item';
 import { DocsFooter } from '../../components/footer';
 import { TableOfContents } from '../../components/navigation/toc';
+import { createReader } from '@keystatic/core/reader';
+import keystaticConfig from '../../../keystatic.config';
 
 const reader = createReader('', keystaticConfig);
 
@@ -41,45 +40,39 @@ export default async function RootLayout({
   }));
 
   return (
-    <div>
-      {/** TOP NAV */}
-      <div className="border-b border-stone-400/20 flex items-center w-full lg:fixed z-20 lg:z-30">
-        <HeaderNav navigationMap={navigationMap} />
-      </div>
+    <>
+      {/* <main className="max-w-7xl min-h-screen mx-auto"> */}
+      <SideNav>
+        {navigationMap?.map(({ groupName, items }) => (
+          <NavGroup key={groupName} title={groupName}>
+            {items.map(({ label, href, title }) => (
+              <NavItem
+                key={href}
+                label={label}
+                href={href}
+                title={title}
+                level="sub"
+              />
+            ))}
+          </NavGroup>
+        ))}
+      </SideNav>
 
-      {/** MAIN */}
-      <main className="max-w-7xl min-h-screen mx-auto">
-        <SideNav>
-          {navigationMap?.map(({ groupName, items }) => (
-            <NavGroup key={groupName} title={groupName}>
-              {items.map(({ label, href, title }) => (
-                <NavItem
-                  key={href}
-                  label={label}
-                  href={href}
-                  title={title}
-                  level="sub"
-                />
-              ))}
-            </NavGroup>
-          ))}
-        </SideNav>
+      {/** CONTENT */}
+      <div className="px-6 flex-1 lg:pl-60 lg:pt-24">
+        <div className="py-10 lg:pl-12">
+          <div className="flex gap-8">
+            {/** INNER CONTENT */}
+            <div className="flex-1">{children}</div>
 
-        {/** CONTENT */}
-        <div className="px-6 flex-1 lg:pl-60 lg:pt-24">
-          <div className="py-10 lg:pl-12">
-            <div className="flex gap-8">
-              {/** INNER CONTENT */}
-              <div className="flex-1">{children}</div>
-
-              {/** TOCs */}
-              <TableOfContents />
-            </div>
-
-            <DocsFooter />
+            {/** TOCs */}
+            <TableOfContents />
           </div>
+
+          <DocsFooter />
         </div>
-      </main>
-    </div>
+      </div>
+      {/* </main> */}
+    </>
   );
 }
